requirements.txt
streamlit==1.31.1
firebase-admin==6.4.0
google-cloud-firestore==2.15.0
streamlit-authenticator==0.2.3
plotly==5.18.0
pandas==2.2.0
python-dotenv==1.0.1
pillow==10.2.0
streamlit-option-menu==0.3.6
protobuf==4.25.2
streamlit-extras==0.4.0
openpyxl==3.1.2
pytz==2024.1 